+++ src/components/FhiButton/FhiButton.tsx
@@ -8,10 +8,11 @@ type Props = {
   outline?: boolean,
   className?: string,
   disabled?: boolean,
+  type?: 'submit' | 'button'
 };
 
 const FhiButton: FC<Props> = ({
-  text, size, secondary, outline, onClick, className, disabled,
+  text, size, secondary, outline, onClick, className, disabled, type,
 }) => {
   let btnClass = `${className} btn`;
   if (size === 'small') {
@@ -35,8 +36,12 @@ const FhiButton: FC<Props> = ({
   } else {
     btnClass += '-primary';
   }
-
-  return <button type="button" className={btnClass} onClick={onClick} disabled={disabled}>{text}</button>;
+  return (
+    // eslint-disable-next-line react/button-has-type
+    <button type={type} className={btnClass} onClick={onClick} disabled={disabled}>
+      {text}
+    </button>
+  );
 };
 
 FhiButton.defaultProps = {
@@ -45,6 +50,7 @@ FhiButton.defaultProps = {
   outline: false,
   className: '',
   disabled: false,
+  type: 'button',
 };
 
 export default FhiButton;