+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 > May 04, 2022
 
+* :boom: **Breaking** Replaced text with children i `FhiButton`
+* :tada: **Enhancement** Added option to set sumbit-type to `FhiButton`
 * :tada: **Enhancement** Added prop for iconLight in `FhiHeader`
 * :tada: **Enhancement** Fixed margin/padding bottom on `FhiHeader`
 * :nut_and_bolt: **New** Added `FhiTabBar`