+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 
 
-## Unreleased
+## 1.0.0
 
-> May 04, 2022
+> May 05, 2022
 
 * :boom: **Breaking** Replaced text with children i `FhiButton`
 * :tada: **Enhancement** Added option to set sumbit-type to `FhiButton`